requirements.txt
# requirements file
# note which revision of python, for example 3.9.6
# in this file, insert all the pip install needs, include revision

# Core dependencies
streamlit>=1.37
pydantic>=2.7
pyyaml>=6.0.2
pandas>=2.0.0
python-dotenv>=1.0.0

# AI/ML dependencies
openai>=1.0.0
crewai>=0.30.0
crewai_tools>=0.4.0